Added custom version support
Added custom version support in form of a local-mc-versions.json located in the storage folder.

You can add a version by adding another element to the array of versions. The element must contain a name, a serviceAPIType and a downloadURL, which can either be "noDownload" or a URL to download the version.